refactor: split dashboard/_build.py into hash and npm modules
Extract hashing/rebuild detection into _build_hash.py and npm
execution/syncing into _build_npm.py. Original _build.py remains
as entry point with re-exports and the maybe_build_ui orchestrator.
